Adding ignite repository.
This is needed, as at the time of commit, we're in a half-product state, where some, but not all, projects define this repository themselves.

For clarity, it's preferred that each project (that needs it) defines the igniterealtime repository in it's own POM. By injecting it through this shared (and obscure) settings file, a degree of implicitness is introduce that will confuse people.